+++ src/profiler.rs
@@ -3,7 +3,6 @@
 use std::convert::TryInto;
 use std::os::raw::c_int;
 
-use backtrace::Frame;
 use nix::sys::signal;
 use once_cell::sync::Lazy;
 use parking_lot::RwLock;
@@ -12,6 +11,7 @@ use smallvec::SmallVec;
 #[cfg(any(target_arch = "x86_64", target_arch = "aarch64"))]
 use findshlibs::{Segment, SharedLibrary, TargetSharedLibrary};
 
+use crate::backtrace::{Frame, Trace, TraceImpl};
 use crate::collector::Collector;
 use crate::error::{Error, Result};
 use crate::frames::UnresolvedFrames;
@@ -54,6 +54,7 @@ impl ProfilerGuardBuilder {
     pub fn frequency(self, frequency: c_int) -> Self {
         Self { frequency, ..self }
     }
+
     #[cfg(all(any(target_arch = "x86_64", target_arch = "aarch64")))]
     pub fn blocklist<T: AsRef<str>>(self, blocklist: &[T]) -> Self {
         let blocklist_segments = {
@@ -124,7 +125,7 @@ pub struct ProfilerGuard<'a> {
 
 fn trigger_lazy() {
     let _ = backtrace::Backtrace::new();
-    let _lock = PROFILER.read();
+    let _profiler = PROFILER.read();
 }
 
 impl ProfilerGuard<'_> {
@@ -181,12 +182,12 @@ fn write_thread_name_fallback(current_thread: libc::pthread_t, name: &mut [libc:
     }
 }
 
-#[cfg(not(all(any(target_os = "linux", target_os = "macos"), target_env = "gnu")))]
+#[cfg(not(any(target_os = "linux", target_os = "macos")))]
 fn write_thread_name(current_thread: libc::pthread_t, name: &mut [libc::c_char]) {
     write_thread_name_fallback(current_thread, name);
 }
 
-#[cfg(all(any(target_os = "linux", target_os = "macos"), target_env = "gnu"))]
+#[cfg(any(target_os = "linux", target_os = "macos"))]
 fn write_thread_name(current_thread: libc::pthread_t, name: &mut [libc::c_char]) {
     let name_ptr = name as *mut [libc::c_char] as *mut libc::c_char;
     let ret = unsafe { libc::pthread_getname_np(current_thread, name_ptr, MAX_THREAD_NAME) };
@@ -208,7 +209,7 @@ extern "C" fn perf_signal_handler(
 ) {
     if let Some(mut guard) = PROFILER.try_write() {
         if let Ok(profiler) = guard.as_mut() {
-            #[cfg(all(any(target_arch = "x86_64", target_arch = "aarch64")))]
+            #[cfg(any(target_arch = "x86_64", target_arch = "aarch64"))]
             if !ucontext.is_null() {
                 let ucontext: *mut libc::ucontext_t = ucontext as *mut libc::ucontext_t;
 
@@ -244,20 +245,23 @@ extern "C" fn perf_signal_handler(
                 }
             }
 
-            let mut bt: SmallVec<[Frame; MAX_DEPTH]> = SmallVec::with_capacity(MAX_DEPTH);
+            let mut bt: SmallVec<[<TraceImpl as Trace>::Frame; MAX_DEPTH]> =
+                SmallVec::with_capacity(MAX_DEPTH);
             let mut index = 0;
+            TraceImpl::trace(ucontext, |frame| {
+                let ip = Frame::ip(frame);
+                if profiler.is_blocklisted(ip) {
+                    return false;
+                }
 
-            unsafe {
-                backtrace::trace_unsynchronized(|frame| {
-                    if index < MAX_DEPTH {
-                        bt.push(frame.clone());
-                        index += 1;
-                        true
-                    } else {
-                        false
-                    }
-                });
-            }
+                if index < MAX_DEPTH {
+                    bt.push(frame.clone());
+                    index += 1;
+                    true
+                } else {
+                    false
+                }
+            });
 
             let current_thread = unsafe { libc::pthread_self() };
             let mut name = [0; MAX_THREAD_NAME];
@@ -349,7 +353,7 @@ impl Profiler {
     // This function has to be AS-safe
     pub fn sample(
         &mut self,
-        backtrace: SmallVec<[Frame; MAX_DEPTH]>,
+        backtrace: SmallVec<[<TraceImpl as Trace>::Frame; MAX_DEPTH]>,
         thread_name: &[u8],
         thread_id: u64,
     ) {